move SpaceTimeAStar to mapf
