+++ w9_pathfinding/src/include/space_time_a_star.h
@@ -5,7 +5,7 @@
 #include "reservation_table.h"
 
 
-class SpaceTimeAStar : public AbsPathFinder {
+class SpaceTimeAStar {
 
     struct Node {
         Node* parent;
